Extract components from pages/event-groups/new.tsx
